Make multiversioning optional
The multiversioning pass currently does two things:
- Clone all functions and create a set of tables to tell the sysimage
  loader where to find the various cloned functions.
- Compress the table of pointers by going from 64 bit pointers to 32 bit
  offsets from the first function of the .text section.

The second optimization is useful, because it cuts down on space
and speed up dynamic loading. Unfortunately relocations of this kind
are not expressible in all object formats and as a result this
scheme does not work if the table needs to describe function pointers
in multiple compilation units.

I'm working on improving the performance of incremental system image
rebuilds which would rely on being able to re-link such system images
and is thus incomptabile with this compression.

There are possible ways, to make it compatible, namely:
- Add a relocation to all the relevant file formats that expresses
  offsets from the start of the section, or,
- Change the multiversion table to be pcrel rather than relative to
  the first function in the table.

The first would require some signifcant coordination with standards
bodies, and both are currently not supported in LLVM.

To make progress on this issue, simply make the multiversion pass
optional and keep the table uncompressed in this case. This wastes
some space and adds a few fractions of a second to the system image
load time, but it should let us proceed on the incremental sysimage
project. If it works well, we can go back and consider the future
of the multiversioning tables.